refactor(server): split server.rs into focused modules (#72)
validate_auth moves to server/auth.rs, the serialization type
pre-checks to server/serialization_checks.rs, and the serde error
mapping to server/serde_errors.rs. The moved bodies are unchanged;
only the four entry points the server calls gain pub(super). Also
points a stale doc reference in types.rs at the new location.
